+++ app/controllers/staff/exports_controller.rb
@@ -41,4 +41,19 @@ class Staff::ExportsController < Staff::BaseController
       end
     end
   end
+
+  def spending_breakdown
+    authorize :export, :show_external_income?
+    fund = Fund.new(params[:fund_id])
+
+    respond_to do |format|
+      format.csv do
+        export = SpendingBreakdown::Export.new(source_fund: fund)
+
+        stream_csv_download(filename: export.filename, headers: export.headers) do |csv|
+          export.rows.each { |row| csv << row }
+        end
+      end
+    end
+  end
 end