+++ spec/policies/export/organisation_policy_spec.rb
@@ -15,6 +15,7 @@ RSpec.describe Export::OrganisationPolicy do
       is_expected.to permit_action(:show_transactions)
       is_expected.to permit_action(:show_xml)
       is_expected.to permit_action(:show_budgets)
+      is_expected.to permit_action(:show_spending_breakdown)
     end
   end
 
@@ -28,6 +29,7 @@ RSpec.describe Export::OrganisationPolicy do
       is_expected.to forbid_action(:show_transactions)
       is_expected.to forbid_action(:show_xml)
       is_expected.to forbid_action(:show_budgets)
+      is_expected.to forbid_action(:show_spending_breakdown)
     end
 
     context "when the user's organisation matches the organisation" do
@@ -40,6 +42,7 @@ RSpec.describe Export::OrganisationPolicy do
         is_expected.to forbid_action(:index)
         is_expected.to forbid_action(:show_transactions)
         is_expected.to forbid_action(:show_xml)
+        is_expected.to permit_action(:show_spending_breakdown)
       end
     end
   end